+++ cpp/src/neighbors/cagra/search_half_dim512_t16.cu
@@ -21,11 +21,11 @@
 #include "raft/neighbors/detail/cagra/search_single_cta.cuh"
 
 namespace raft::neighbors::experimental::cagra::detail::single_cta_search {
-template struct search<16, 512, half, float, uint32_t>;
+template struct search<16, 512, half, uint32_t, float>;
 }
 namespace raft::neighbors::experimental::cagra::detail::multi_cta_search {
-template struct search<16, 512, half, float, uint32_t>;
+template struct search<16, 512, half, uint32_t, float>;
 }
 namespace raft::neighbors::experimental::cagra::detail::multi_kernel_search {
-template struct search<16, 512, half, float, uint32_t>;
+template struct search<16, 512, half, uint32_t, float>;
 }
